organise tests, add permissions tests
