+++ pallets/fellowship/src/lib.rs
@@ -403,5 +403,6 @@ pub mod pallet {
     pub enum Permission {  
         AddToShortlist,
         RemoveFromShortlist,
+        None,
     }
 }
