+++ pallets/fellowship/src/mock.rs
@@ -77,7 +77,7 @@ impl pallet_fellowship::Config for Test {
     type MembershipDeposit = MembershipDeposit;
     type DepositCurrencyId = DepositCurrencyId;
     type SlashAccount = SlashAccount;
-    type Permissions = Test;
+    type Permissions = crate::impls::VetterAndFreelancerAllPermissions;
     type WeightInfo = ();
 }
 
@@ -127,16 +127,6 @@ pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
         let _ = Tokens::deposit(CurrencyId::Native, &TREASURY, initial_balance);
     });
     ext
-}
-
-impl crate::traits::FellowshipPermissions<crate::Role, crate::Permission> for Test {
-    fn has_permission(role: Role, permission: Permission) -> bool {
-        true
-    }
-
-    fn get_permisisons(role: Role) -> Vec<Permission> {
-        Vec::new()
-    }
 }
 
 use frame_support::pallet_prelude::Weight;
