+++ pallets/fellowship/src/tests/pallet_tests.rs
@@ -1,129 +1,4 @@
-use crate::impls::*;
-use crate::traits::*;
-use crate::*;
-use crate::{mock::*, Error, Event, FellowToVetter, Role, Roles};
-use common_traits::MaybeConvert;
-use common_types::CurrencyId;
-use frame_support::{
-    assert_noop, assert_ok, once_cell::sync::Lazy, traits::Hooks, BoundedBTreeMap,
-};
-use frame_system::Pallet as System;
-use orml_tokens::Error as TokensError;
-use orml_traits::{MultiCurrency, MultiReservableCurrency};
-use sp_arithmetic::traits::One;
-use sp_core::sr25519::Public;
-use sp_runtime::{traits::BadOrigin, DispatchError, Saturating};
-use sp_std::vec;
-
-// Saves a bit of typing.
-pub(crate) static DEP_CURRENCY: Lazy<CurrencyId> =
-    Lazy::new(<Test as Config>::DepositCurrencyId::get);
-fn add_to_fellowship_take_deposit(
-    who: &AccountIdOf<Test>,
-    role: Role,
-    rank: Rank,
-    vetter: Option<&VetterIdOf<Test>>,
-) -> Result<(), DispatchError> {
-    <Fellowship as FellowshipHandle<AccountIdOf<Test>>>::add_to_fellowship(
-        who, role, rank, vetter, true,
-    );
-    Ok(())
-}
-
-fn revoke_fellowship(who: &AccountIdOf<Test>, slash_deposit: bool) -> Result<(), DispatchError> {
-    <Fellowship as FellowshipHandle<AccountIdOf<Test>>>::revoke_fellowship(who, slash_deposit)
-}
-
-pub fn run_to_block<T: Config>(n: T::BlockNumber)
-where
-    T::BlockNumber: Into<u64>,
-{
-    loop {
-        let mut block: T::BlockNumber = frame_system::Pallet::<T>::block_number();
-        if block >= n {
-            break;
-        }
-        block = block.saturating_add(<T::BlockNumber as One>::one());
-        frame_system::Pallet::<T>::set_block_number(block);
-        frame_system::Pallet::<T>::on_initialize(block);
-        Fellowship::on_initialize(block.into());
-    }
-}
-
-#[test]
-fn ensure_role_in_works() {
-    new_test_ext().execute_with(|| {
-        Roles::<Test>::insert(*ALICE, (Role::Vetter, 10));
-        Roles::<Test>::insert(*BOB, (Role::Freelancer, 10));
-
-        assert_ok!(EnsureFellowshipRole::<Test>::ensure_role_in(
-            &ALICE,
-            vec![Role::Vetter, Role::Freelancer],
-            None
-        ));
-        assert_ok!(EnsureFellowshipRole::<Test>::ensure_role_in(
-            &BOB,
-            vec![Role::Vetter, Role::Freelancer],
-            None
-        ));
-        assert!(
-            EnsureFellowshipRole::<Test>::ensure_role_in(&BOB, vec![Role::Vetter], None).is_err(),
-            "BOB is not of this Role."
-        );
-        assert!(
-            EnsureFellowshipRole::<Test>::ensure_role_in(&ALICE, vec![Role::Freelancer], None)
-                .is_err(),
-            "ALICE is not of this Role."
-        );
-    });
-}
-
-#[test]
-fn ensure_role_in_works_with_rank() {
-    new_test_ext().execute_with(|| {
-        Roles::<Test>::insert(*ALICE, (Role::Vetter, 10));
-        assert_ok!(EnsureFellowshipRole::<Test>::ensure_role_in(
-            &ALICE,
-            vec![Role::Vetter],
-            Some(vec![10, 9])
-        ));
-
-        assert_noop!(
-            EnsureFellowshipRole::<Test>::ensure_role_in(&ALICE, vec![Role::Vetter], Some(vec![9])),
-            BadOrigin
-        );
-    });
-}
-
-#[test]
-fn ensure_role_works() {
-    new_test_ext().execute_with(|| {
-        Roles::<Test>::insert(*ALICE, (Role::Vetter, 0));
-        assert_ok!(EnsureFellowshipRole::<Test>::ensure_role(
-            &ALICE,
-            Role::Vetter,
-            None
-        ));
-        assert!(EnsureFellowshipRole::<Test>::ensure_role(&ALICE, Role::Freelancer, None).is_err());
-    });
-}
-
-#[test]
-fn ensure_role_works_with_rank() {
-    new_test_ext().execute_with(|| {
-        Roles::<Test>::insert(*ALICE, (Role::Vetter, 10));
-        assert_ok!(EnsureFellowshipRole::<Test>::ensure_role(
-            &ALICE,
-            Role::Vetter,
-            Some(10)
-        ));
-
-        assert_noop!(
-            EnsureFellowshipRole::<Test>::ensure_role(&ALICE, Role::Vetter, Some(9)),
-            BadOrigin
-        );
-    });
-}
+use super::*;
 
 #[test]
 fn freelancer_to_vetter_works() {
